requirements.txt
Flask==2.1.3
PyJWT==2.4.0
requests==2.28.0
